upgraded to sdl3

+++ Input.h
@@ -1,18 +1,18 @@
 #pragma once
 #include <string>
 #include <utility>
-#include <SDL.h>
-class Input {
+#include <SDL3/SDL.h>
+class Input
+{
     bool mouseButtonDown_Left = false;
     bool mouseButtonDown_Right = false;
     bool mouseButtonDown_Middle = false;
 
 public:
-    std::pair<int, int> currentMousePosition;
-    std::pair<int, int> prevMousePosition;
+    std::pair<float, float> currentMousePosition;
+    std::pair<float, float> prevMousePosition;
 
-    void MouseButtonDown(SDL_MouseButtonEvent& b);
-    void MouseButtonUp(SDL_MouseButtonEvent& b);
+    void MouseButtonDown(SDL_MouseButtonEvent &b);
+    void MouseButtonUp(SDL_MouseButtonEvent &b);
     bool GetMouseButtonDown(std::string button);
 };
-
